+++ versions/v20/projects/sdcorejs-angular/components/card/src/card.constants.ts
@@ -1,0 +1,3 @@
+import type { Color as SdColor } from '@sdcorejs/utils/models';
+
+export const SD_CARD_DEFAULT_COLOR: SdColor = 'primary';